requirements.txt
flask>=2.0.2
pyjwt>=1.5.3
pyimgur>=0.6.0